#  Faster StringDictionaryBuilder (~60% faster) (#1851) 